More to change TimeTrigger to Trigger.

+++ src/AutoHome.Core/Entities/Trigger.cs
@@ -5,5 +5,5 @@ public class Trigger : IEntity
     public Guid Id { get; set; }
     public Guid DeviceId { get; set; }
     public string Name { get; set; } = string.Empty;
-    public long Interval { get; set; }
+    public double Interval { get; set; }
 }
